move cypress folder into app
