+++ app/tsconfig.json
@@ -15,6 +15,6 @@
         "jsx": "preserve",
         "incremental": true
     },
-    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "**/*.d.ts"],
-    "exclude": ["node_modules"]
+    "include": ["next-env.d.ts", "src/", "public/", "pages/", "./index.d.ts"],
+    "exclude": ["node_modules", "./cypress"]
 }
